requirements.txt
annotated-types==0.7.0
anthropic==0.67.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cobble==0.1.4
colorama==0.4.6
comm==0.2.3
cryptography==45.0.7
curl_cffi==0.13.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
frozendict==2.4.6
google_search_results==2.4.2
googleapis-common-protos==1.69.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.1
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathvalidate==3.2.3
pdfminer==20191125
pdfminer.six==20250327
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotext==5.3.2
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pyzmq==27.1.0
questionary==2.1.1
regex==2025.9.1
requests==2.32.5
rich==14.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
tabulate==0.9.0
termcolor==3.1.0
tiktoken==0.11.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
xlsxwriter==3.2.8
yfinance==0.2.65
youtube-transcript-api==0.6.3
